Move restart-related tests to their own file
